fix: Move TimestampSummary to the only place it's used
