+++ mutable_batch/Cargo.toml
@@ -10,6 +10,7 @@ arrow_util = { path = "../arrow_util" }
 chrono = { version = "0.4", default-features = false }
 data_types = { path = "../data_types" }
 data_types2 = { path = "../data_types2" }
+iox_time = { path = "../iox_time" }
 schema = { path = "../schema" }
 snafu = "0.7"
 hashbrown = "0.12"
